shims/unix: split general FD management from FS access; make a place for sockets
